+++ plugins/file-uploader-wakamiti-plugin/src/main/java/es/iti/wakamiti/fileuploader/StandardOutputFilesUploader.java
@@ -4,7 +4,7 @@ import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.event.Event;
 import es.iti.wakamiti.api.extensions.EventObserver;
 
-@Extension(provider =  "es.iti.wakamiti", name = "standard-output-files-uploader", version = "2.4")
+@Extension(provider =  "es.iti.wakamiti", name = "standard-output-files-uploader", version = "2.5")
 public class StandardOutputFilesUploader extends AbstractFilesUploader implements EventObserver {