+++ plugins/html-report-wakamiti-plugin/src/main/java/es/iti/wakamiti/report/html/HtmlReportGeneratorConfig.java
@@ -16,7 +16,7 @@ import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.extensions.ConfigContributor;
 
 
-@Extension(provider =  "es.iti.wakamiti", name = "html-report-config", version = "2.4",
+@Extension(provider =  "es.iti.wakamiti", name = "html-report-config", version = "2.5",
     extensionPoint =  "es.iti.wakamiti.api.extensions.ConfigContributor")
 public class HtmlReportGeneratorConfig implements ConfigContributor<HtmlReportGenerator> {