+++ wakamiti-api/src/test/java/es/iti/wakamiti/api/util/JsonUtilsTest.java
@@ -34,9 +34,7 @@ public class JsonUtilsTest {
     @Test
     public void testJsonWhenStringWithSuccess() {
         JsonNode obj = JsonUtils.json(json);
-
-        assertThat(obj).isNotNull();
-        assertThat(obj.toString()).isEqualTo(json);
+        assertThat(obj).hasToString(json);
     }
 
     @Test(expected = RuntimeException.class)
@@ -47,9 +45,7 @@ public class JsonUtilsTest {
     @Test
     public void testJsonWhenStreamWithSuccess() {
         JsonNode obj = JsonUtils.json(new ByteArrayInputStream(json.getBytes()));
-
-        assertThat(obj).isNotNull();
-        assertThat(obj.toString()).isEqualTo(json);
+        assertThat(obj).hasToString(json);
     }
 
     @Test(expected = RuntimeException.class)