+++ wakamiti-engine/wakamiti-core/src/test/java/es/iti/wakamiti/test/gherkin/TestFailureStep.java
@@ -6,7 +6,6 @@
 package es.iti.wakamiti.test.gherkin;
 
 
-import es.iti.wakamiti.api.WakamitiConfiguration;
 import es.iti.wakamiti.api.plan.*;
 import es.iti.wakamiti.api.util.WakamitiLogger;
 import es.iti.wakamiti.core.JsonPlanSerializer;
@@ -22,6 +21,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import static es.iti.wakamiti.api.WakamitiConfiguration.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 
@@ -33,15 +33,16 @@ public class TestFailureStep {
     public void testInvalidStep() throws IOException {
 
         Map<String, String> properties = new HashMap<>();
-        properties.put(WakamitiConfiguration.RESOURCE_TYPES, GherkinResourceType.NAME);
+        properties.put(RESOURCE_TYPES, GherkinResourceType.NAME);
         properties.put(
-                WakamitiConfiguration.RESOURCE_PATH,
+                RESOURCE_PATH,
                 "src/test/resources/features/failure/failure.feature"
         );
         properties.put(
-                WakamitiConfiguration.NON_REGISTERED_STEP_PROVIDERS,
+                NON_REGISTERED_STEP_PROVIDERS,
                 "es.iti.wakamiti.test.gherkin.WakamitiSteps"
         );
+        properties.put(OUTPUT_FILE_PATH, "target/wakamiti.json");
         Configuration configuration = Wakamiti.defaultConfiguration()
                 .appendFromMap(properties);
         PlanNode plan = Wakamiti.instance().createPlanFromConfiguration(configuration);