+++ plugins/azure-wakamiti-plugin/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
 
+[Unreleased]
+
+### Changed
+- Upgrade dependencies: `wakamiti-plugin-starter`
+
+
 ## [1.5.0] - 2023-10-26
 
 ### Added