+++ plugins/azure-wakamiti-plugin/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [1.8.0] - 2024-04-16
+
+### Changed
+- Upgrade dependencies: `wakamiti-plugin-starter`
+
 
 ## [1.7.0] - 2024-03-22